pyproject.toml
[project]
name = "adk-oauth-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.4",
    "google-adk>=1.15.1",
    "google-cloud-secret-manager>=2.24.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "starlette>=0.48.0",
    "tink[gcpkms]>=1.12.0",
    "uvicorn>=0.37.0",
]
